need to write tests
The bare bones parser is done, but tests are required.

The error handling feels _wrong_ lalrpop doesn't document if the
resulting span value (usize) are the character count, or byte
index of the char. I think it is the later (if this is documented
I'm not sure where).

If it is giving us indexes it may work, we will see.