trying to do linkages and failing.
So the hashedexpression stuff isnt doing "enough" it should be
handling identifiers as well as expressions, but it isnt.

The Identifier think in linkage needs to end up in cfgbuilder also
since CFG building will error on lack of returns/analysis i should
write the interior type to get rid of it.

The biggest todo remains:
- Bring the old parsing back so we can write full tests for namespace
  and cfg builders.