Same status as before
The identifier change is needed in the CFGBuilder package pretty
badily that is going to be a bit bigger change then I'd orginally
liked.

Part of what complicated it is that in the namespace package we were
erasing details about _where_ a variable came from, for example if it
was externally defined (a constant, or function argument) then its
source was erased.

While our namespace is flat (function args, constants, and vars all
share a global pool) this still creates a challenge as determining
where a variable's values came from was a little impossible.