attempted to add serde, dont try that again.
I wanted to add serde as it sounded like a thing to do. As then any
stage of the compile process could be dumped to disk and mulled over
which would make testing & verification a lot easier.

ALAS! serde doesn't appear to be in a mood where it can reasonably
handle its own fucking examples. Disappointing to be honest. I've hit
this before and I really don't want to serialize all this junk together
by hand.