some real progress
Re-enabled the full AST builder which means we should bring this
branch up to parity with master as it is now re-anbling all the
features master had.

Got some namespace tests written, and found a bug which is exciting
There will have to be better handling of the standard library.