Minor screw up.
Constant folding is a work in progess, part of the issue is dealing
with functions. They need to be flatted for analysis, while capturing
their environment (kind of).

This means there needs to be some notion of a "stack" when querying
values within CFGBuilder to determine how deep we've currently
recursed when invoking functions.

TODO:
- Functions need to include all external constants in their variable pools.
- Function invocations require contextual information.
- cfgbuilder/collection needs to respect this when querying.