Well I kind of got function querying worked out
The problem is I mostly haven't finished the part of the tests where
I walk the function args but up to their defination. There should be
a short hand for this, but it doesn't exist, and is messy.

I know I'm doing something wrong because the borrow checker is
whining.

TODO:
* finish cfg tests
* add a 3^rd stack to the CallStack for tracking function hashes&
  argument hashes
* create a collection for storing all these InlinedExpression values
  bonus points for dynamic programming.

go to bed