working on constant folding
Constant folding is started. It is not as painful as I thought
it might be, the type checker _should_ be catching the incorrect
errors so we'll see in testing if that works.

The Dynamic Programming aspect as added, and slots in nicely to the
current approach so for complex constant expressions we might be more
consistent.

The hard part now will be serializing all of this into a valid
rustc file, and attempting to compile the output. Everything but
stdlib is being inlined (this needs to be tested) but as that is
done, we can likely do a 1:1 expr -> vec (for collections)

This will likely mean we need to do a secondary pass for constant
propagation.