Progress
Started on the graph ordering pass. I went a little overboard with
the late binding on the expression stuff, but I think it is okay
because the `hash` identifiers are now special values.

Replaced most instances of `panic` and `unreachable` with the new
`_unreachable_panic` which can be configured into an unsafe
"this branch will never occur" hint. The goal of this is to reduce
branches which preform pointless matches on conditional values,
while I pass around `Option<T>` A LOT that is mostly for convience,
we shouldn't ever get `None`. This allows for a user to enable
"drop_unreachable_branches" which will convert the various
"panic" and "unreachable" invocations into UB, allowing for branches
to be optimized out of the final code. This should improve the final
runtime considerably.

TODO:
- Ordering pass complete
- Ordering pass tests
- Scalar vs Vec pass
- Scalar vs Vec tests
- Figure out how to actually execute these programs?